Refactor patron Strategy
